# documentation/swagger.yaml
openapi: 3.0.3
info:
  title: Business Agent API
  version: '1.0.0'
  description: >-
    API surface for triggering agents from the Business Agent console. The
    primary entry point is a POST request that queues work, and a GET endpoint
    is available for streaming back logs and final results.
servers:
  - url: http://localhost:3000
paths:
  /api/agents/trigger:
    post:
      summary: Trigger an agent
      description: >-
        Queues execution for the given agent identifier without holding the HTTP
        connection open. The endpoint proxies payloads to the `agents` package or
        returns a stubbed response when the runner is unavailable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agentId
              properties:
                agentId:
                  type: string
                  description: Name of the agent to trigger.
                  example: CEO
                input:
                  description: Arbitrary payload forwarded to the agent.
                  nullable: true
      responses:
        '202':
          description: Agent execution queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  agentId:
                    type: string
                  runId:
                    type: string
                    description: Identifier for retrieving run logs.
                  status:
                    type: string
                    description: Current state of the run.
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Server-side failure to trigger the agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/agents/{runId}:
    get:
      summary: Get an agent run snapshot
      description: Returns current status, logs, and any result payload for a queued or completed run.
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: string
          description: Identifier returned from the trigger endpoint.
      responses:
        '200':
          description: Run details
          content:
            application/json:
              schema:
                type: object
                properties:
                  runId:
                    type: string
                  agentId:
                    type: string
                  status:
                    type: string
                    enum: [queued, running, completed, failed]
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        timestamp:
                          type: string
                          format: date-time
                        message:
                          type: string
                        payload:
                          type: object
                          nullable: true
                  result:
                    description: Raw result from the agent runner when available.
                    type: object
                    nullable: true
        '404':
          description: Run not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/todos:
    post:
      summary: Create a todo entry
      description: Writes the todo body to disk and records metadata in a SQLite database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - body
              properties:
                title:
                  type: string
                  description: Title for the todo entry.
                body:
                  type: string
                  description: Detailed content saved to the task file.
      responses:
        '201':
          description: Todo persisted
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        '400':
          description: Missing parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Failure saving the todo
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    get:
      summary: List todos
      description: Returns saved todo metadata without loading the file contents.
      responses:
        '200':
          description: Todo list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
        '500':
          description: Failure loading todos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/todos/{todoId}:
    get:
      summary: Load todo from file
      description: Retrieves todo metadata and the stored file contents using the SQLite index.
      parameters:
        - in: path
          name: todoId
          required: true
          schema:
            type: integer
          description: Identifier returned by the create endpoint.
      responses:
        '200':
          description: Todo content
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  body:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Failure loading todo
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string